refactor: split codegen CLI
