+++ cmd/codegen/generator/client.go
@@ -1,0 +1,2 @@
+package generator
+
