+++ cmd/codegen/generator/go/templates/modules.go
@@ -30,19 +30,19 @@ const (
 )
 
 func (funcs goTemplateFuncs) isModuleCode() bool {
-	return funcs.cfg.ModuleName != ""
+	return funcs.cfg.ModuleConfig != nil
 }
 
 func (funcs goTemplateFuncs) isStandaloneClient() bool {
-	return funcs.cfg.ClientOnly
+	return funcs.cfg.ClientConfig != nil
 }
 
 func (funcs goTemplateFuncs) Dependencies() []generator.ModuleSourceDependencies {
-	return funcs.cfg.ModuleDependencies
+	return funcs.cfg.ClientConfig.ModuleDependencies
 }
 
 func (funcs goTemplateFuncs) HasLocalDependencies() bool {
-	for _, dep := range funcs.cfg.ModuleDependencies {
+	for _, dep := range funcs.cfg.ClientConfig.ModuleDependencies {
 		if dep.Kind == "LOCAL_SOURCE" {
 			return true
 		}
@@ -56,7 +56,7 @@ func (funcs goTemplateFuncs) moduleRelPath(path string) string {
 		// path to the root of this module (since we're probably in internal/dagger/)
 		"../..",
 		// path from the module root to the context directory
-		funcs.cfg.ModuleParentPath,
+		funcs.cfg.ModuleConfig.ModuleParentPath,
 		// path from the context directory to the desired path
 		path,
 	)
@@ -100,7 +100,7 @@ func (funcs goTemplateFuncs) moduleMainSrc() (string, error) { //nolint: gocyclo
 		pkg:        funcs.modulePkg,
 		fset:       funcs.moduleFset,
 		schema:     funcs.schema,
-		moduleName: funcs.cfg.ModuleName,
+		moduleName: funcs.cfg.ModuleConfig.ModuleName,
 
 		methods: make(map[string][]method),
 	}
