+++ cmd/codegen/generator/module.go
@@ -1,0 +1,2 @@
+package generator
+
