+++ cmd/codegen/generator/schema.go
@@ -1,6 +1,8 @@
 package generator
 
-import "github.com/dagger/dagger/cmd/codegen/introspection"
+import (
+	"github.com/dagger/dagger/cmd/codegen/introspection"
+)
 
 var _schema *introspection.Schema
 
@@ -10,4 +12,13 @@ func SetSchema(schema *introspection.Schema) {
 
 func GetSchema() *introspection.Schema {
 	return _schema
+}
+
+// SetSchemaParents sets all the parents for the fields.
+func SetSchemaParents(schema *introspection.Schema) {
+	for _, t := range schema.Types {
+		for _, f := range t.Fields {
+			f.ParentObject = t
+		}
+	}
 }
