+++ cmd/codegen/generator/typescript/templates/functions_test.go
@@ -21,7 +21,7 @@ func init() {
 		panic(err)
 	}
 
-	currentSchema, _, err = generator.Introspect(ctx, c)
+	currentSchema, _, err = introspection.Introspect(ctx, c)
 	if err != nil {
 		panic(err)
 	}
